test: tests in the Radio component
